Add 'Owner' field to 'Trail' entity
This update introduces an 'Owner' field to the 'Trail' entity as a string. 'Owner' field is integrated in the database migration, to the entity model, and adjusted in the client-side components. This change enforces that each trail has an owner and helps to provide trail management according to the owner.